refactor: Colocate fetch-actor-details schemas and response builders
Move tool-specific items from src/utils/actor_details.ts to
src/tools/core/fetch_actor_details_common.ts — matching the structure
of src/tools/core/search_actors_common.ts:

  - actorDetailsOutputOptionsSchema
  - actorDetailsOutputDefaults
  - ResolvedOutputOptions
  - resolveOutputOptions
  - buildActorNotFoundResponse
  - buildActorDetailsTextResponse

utils/actor_details.ts now holds only shared data-fetching and
formatting utilities (fetchActorDetails, getMcpToolsMessage,
processActorDetailsForResponse, resolveReadmeContent,
buildCardOptions, typeObjectToString, ActorDetailsResult).

Export typeObjectToString since buildActorDetailsTextResponse (now
in tools/core) still uses it.
